Client filter. Repository exception custom. Documentation update.

+++ app/Services/AdService.php
@@ -9,10 +9,15 @@ namespace App\Services;
 
 
 use App\Ads;
+use App\Criterias\ClientCriteria;
+use App\Exceptions\Traits\ClientTrait;
 use App\Files;
 use App\Repositories\AdRepository;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class AdService {
+
+	use ClientTrait;
 
     private $ad;
     private $fileService;
@@ -24,27 +29,28 @@ class AdService {
     }
 
     function all() {
-    	return $this->ad->all();
+    	$clientId = $this->getClientId();
+    	return $this->ad->pushCriteria(new ClientCriteria($clientId))->all();
 	}
 
     function find($id) {
-        return $this->ad->find($id);
+        return $this->ad->pushCriteria(new ClientCriteria($this->getClientId()))->find($id);
     }
 
     function paginate($perPage) {
     	$perPage = $perPage ?? 5;
-		$data = $this->ad->paginate($perPage);
+    	$clientId = $this->getClientId();
+		$data = $this->ad->pushCriteria(new ClientCriteria($clientId))->paginate($perPage);
 		return $data;
 	}
 
     function create($data) {
         /** @var Files $file */
 //        $file = factory(Files::class)->create();
-		$curato = app('curato');// Initial at the ApiAuthenticationMiddleware
 		$file = $this->fileService->store('ads');// Storing to directory: public/ads
         $ad = new Ads();
 //        $ad->client_id = 1;
-		$ad->client_id = $curato->client->id;
+		$ad->client_id = $this->getClientId();
         $ad->file_id = $file->id;
 		$ad->fill($data);
 		$ad->save();
@@ -56,6 +62,10 @@ class AdService {
     	/** @var Ads $ad */
     	$ad = $this->ad->skipPresenter()->find($id);
 
+    	if ($ad->client_id != $this->getClientId()) {
+    		throw new NotFoundHttpException('Resource not found');
+		}
+
     	$ad->fill($data);
     	$ad->save();
 
@@ -64,6 +74,11 @@ class AdService {
 	}
 
 	function delete($id) {
+    	/** @var Ads $ad */
+    	$ad = $this->ad->skipPresenter()->find($id);
+    	if ($ad->client_id != $this->getClientId()) {
+    		throw new NotFoundHttpException('Resource not found');
+		}
     	return $this->ad->delete($id);
 	}
 }
